telegram-stock-bot/requirements.txt
python-telegram-bot==20.7
yfinance==0.2.40
pandas==2.1.4
python-dotenv==1.0.0
